Original Compare two sheets.
The purpose of this vba is to compare two sheets and establish a common column in order to match on. Once it runs it will highlight in Red something that does not match. Finally, it provides a result sheet to provide statistics and the list the items that do not match and where they are.